refactor: update SkillCard component to use utility functions and improve styling
