+++ portfolio/src/components/Portfolio/toolbox/index.tsx
@@ -6,7 +6,7 @@ const Toolbox = () => {
   return (
     <>
       <ToolboxHeader />
-      <div className='relative mt-12 mb-16 min-h-[300px] sm:min-h-[400px] flex justify-center'>
+      <div className='relative mt-12 mb-34 md:mb-2 min-h-[300px] sm:min-h-[400px] flex justify-center'>
         {skillsData.map((category, index) => (
           <SkillCard key={category.title} category={category} index={index} />
         ))}
@@ -14,5 +14,4 @@ const Toolbox = () => {
     </>
   )
 }
-
 export default Toolbox
